+++ src/FileData/Curiosity.FileDataReaderWriters/Style/TextAlignment.cs
@@ -1,0 +1,9 @@
+﻿namespace Curiosity.FileDataReaderWriters.Style;
+
+public enum TextAlignment
+{
+    Left = 0,
+    Center = 1,
+    Right = 2,
+    Justify = 3
+}